SURPRISE! Let's switch embedded dbs.
Now using kv + sled instead of sqlite, because what's a side
project for if not for excursions into unknown territory?
Also because I was not taking advantage of anything a relational
database could do for me, and sqlite made me miss postgres.

Advantage: serde serialization is now doing double-duty.

There's some cleanup work to be done BUT I can now move on to the
next problem.