feature: store & use etags so we can update when we need to
Sadly, we still get dinged an API request for a cache hit, so
this isn't worth all that much, but it's the principle.

Also, did the first of several refactorings to pull repeated
code up into parametrized functions so things get implemented
once, and one hopes properly.

Sadly the http fetching functions got worse with code repetition
and not better, but eventually the patterns will become obvious
to me.