Refactoring and readability cleanup commences
We start by pulling the k/v store global into a once_cell,
which lets us remove repeated noise from the API since we
don't have to pass it around like a baton any more. This
is only partially accomplished.

The more notable change in this commit is that many command
implementations are now in their own files in `src/commands`.
This both makes main more readable and highlights the repeated
cruft in the APIs.